/settings was added

+++ bot/commands/login/menu.py
@@ -1,5 +1,4 @@
 from aiogram.types import CallbackQuery
-from aiogram.types import Message
 
 from bot import dispatcher
 from bot import students
@@ -11,39 +10,30 @@ from bot.shared.helpers import top_notification
 from bot.shared.commands import Commands
 
 
-@dispatcher.message_handler(
-    lambda message:
-        students[message.chat.id].guard.text == Commands.START.value or
-        students[message.chat.id].guard.text is None,
-    commands=[ Commands.LOGIN.value ]
+@dispatcher.callback_query_handler(
+    lambda callback: (
+            students[callback.message.chat.id].guard.text == Commands.START.value or
+            students[callback.message.chat.id].guard.text == Commands.SETTINGS.value
+        ) and callback.data == Commands.LOGIN.value
 )
 @metrics.increment(Commands.LOGIN)
-async def login_on_command(message: Message):
-    await message.answer(
+@top_notification
+async def login(callback: CallbackQuery):
+    # Cleanning the chat
+    await callback.message.delete()
+    if students[callback.message.chat.id].guard.text == Commands.START.value: await students[callback.message.chat.id].guard.message.delete()
+    
+    await callback.message.answer(
         text=(
             "{warning}"
             "Студенческий билет и зачётка имеют одинаковый номер😉\n\n"
             "Выбери желаемый путь настройки:"
         ).format(
             # Showing the warning to the old users
-            warning="Все текущие данные, включая *заметки* и *изменённое расписание*, будут стёрты.\n\n" if students[message.chat.id].is_setup else ""
+            warning="Все текущие данные, включая *заметки* и *изменённое расписание*, будут стёрты.\n\n" if students[callback.message.chat.id].is_setup else ""
         ),
         parse_mode="markdown",
-        reply_markup=login_way_chooser(is_old=students[message.chat.id].is_setup)
+        reply_markup=login_way_chooser(is_old=students[callback.message.chat.id].is_setup)
     )
     
-    students[message.chat.id].guard.text = Commands.LOGIN.value
-
-@dispatcher.callback_query_handler(
-    lambda callback:
-        students[callback.message.chat.id].guard.text == Commands.START.value and
-        callback.data == Commands.LOGIN.value
-)
-@metrics.increment(Commands.LOGIN)
-@top_notification
-async def login_on_callback(callback: CallbackQuery):
-    # Cleanning the chat
-    await callback.message.delete()
-    await students[callback.message.chat.id].guard.message.delete()
-    
-    await login_on_command(callback.message)
+    students[callback.message.chat.id].guard.text = Commands.LOGIN.value
